Remove CalibrationFile from config.
