+++ api/models/config_keys.py
@@ -91,7 +91,6 @@ class SourceConfigDTO(BaseModel):
     violationThreshold: Optional[int] = Field(0, example=100)
     image: Optional[str] = Field("", example='Base64 image')
     distMethod: Optional[str] = Field("", example='CenterPointsDistance')
-    calibrationFile: Optional[str] = Field("", example='/repo/data/processor/static/data/cam1/homography_matrix/h_inverse.txt')
     dailyReport: Optional[bool] = Field(False, example=True)
