+++ api/tests/data/config-sample.ini
@@ -21,7 +21,6 @@ Name = Front
 Id = cam1
 Emails =
 DistMethod = distance_method_this_is_test_sample_config
-CalibrationFile = /repo/data/processor/static/data/cam1/homography_matrix/h_inverse.txt
 
 [Source_1]
 VideoPath = /repo/data/softbio_vid.mp4
@@ -30,7 +29,6 @@ Name = Backyard
 Id = cam2
 Emails =
 DistMethod =
-CalibrationFile = /repo/data/processor/static/data/cam2/homography_matrix/h_inverse.txt
 
 [PostProcessor]
 MaxTrackFrame = 5
