+++ config-x86.ini
@@ -32,7 +32,6 @@ Id = default
 Emails =
 NotifyEveryMinutes = 0
 ViolationThreshold = 60
-CalibrationFile = /repo/data/processor/static/data/cam1/homography_matrix/h_inverse.txt
 ; Distance measurement method:
 ; - CalibratedDistance: calculate the distance with 3-d transformed points, note that by choosing this method you should specify the inverse calibration matrix of your environment.
 ; - CenterPointsDistance: compare center of pedestrian boxes together
